feat: OpenDDE predicts on a bfloat16 trunk by default
It shipped float32 because upstream does, under the rule that memory alone
does not move a default away from what upstream runs. It is not memory alone.
At 1,003 tokens, five samples, two runs per arm: peak 42.96 -> 19.18 GiB, wall
280.4 -> 174.4 s, best ranking_score unchanged, and a median CA RMSD of 0.0301
A against the 0.0142 A floor from running either arm twice -- about twice the
noise, with one sample of five moving 0.6140 A, so it is a real difference
rather than none. At 1,531 tokens float32 asks for 86.7 GiB and dies where
bfloat16 finishes at 44.2 GiB.

The benchmark pins float32 for OpenDDE so its row stays a comparison: the
driver passes no options, so the flip would have made the FoldJAX column
bfloat16 against upstream's float32, silently, because a default is the one
thing nobody passes. The pin covers the warm-up too, so the compiled program
and the measured program stay the same.

Two OpenDDE flows had never exercised the shipped weight cast -- their weights
double was object(), which only works while float32 casts nothing.

+++ src/foldjax/models/opendde/cli/predict.py
@@ -456,10 +456,29 @@ def main(
     parser.add_argument(
         "--trunk-dtype",
         choices=("bf16", "fp32"),
-        default="fp32",
+        default="bf16",
         help="element width of the embedder and both trunks; the diffusion "
         "sampler and the output heads stay FP32 either way",
     )
+    # This default is FoldJAX's, not upstream's. Upstream OpenDDE ships fp32
+    # (`opendde/config/model_base.py:37`) and this port shipped fp32 with it
+    # until 2026-08-28, on the rule that memory alone does not move a default
+    # away from what upstream runs. It is no longer memory alone. Measured at
+    # 1,003 tokens, five samples, two runs per arm:
+    #
+    #     peak     42.96 -> 19.18 GiB   (-55.4%)
+    #     wall    280.4  -> 174.4 s     (-37.8%)
+    #     best ranking_score  0.1926 -> 0.1926
+    #     CA RMSD  median 0.0301 A, against a 0.0142 A floor from running the
+    #              same arm twice -- about 2x the noise, and one of the five
+    #              samples moved 0.6140 A, so it is a real difference rather
+    #              than none at all.
+    #
+    # Smaller and faster and inside its own sampling spread, at three sizes:
+    # 1,354 (335 s / 34.7 GiB against 529 s / 77.8 GiB) and 1,531, where fp32
+    # asks for 86.7 GiB and dies while bf16 finishes at 44.2 GiB. fp32 stays
+    # one flag away, and `bench/` pins it for the upstream comparison, because
+    # a comparison run at two precisions is not a comparison.
     parser.add_argument("--include-raw", action="store_true")
     parser.add_argument(
         "--representations-dir",
